pyproject.toml
[tool.poetry]
name = "discord-bots"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
ruff = "0.8.2"
mypy = "1.13.0"
discord-py = {version = "2.4.0", extras = ["voice"]}
python-dotenv = "1.0.1"


[tool.ruff]
indent-width = 4
line-length = 99
target-version = "py312"

[tool.ruff.lint]
ignore = [
    "D100", # Missing docstring in public module
    "D400", # First line should end with a period
]
select = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.format]
indent-style = "space"
quote-style = "double"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
